pyproject.toml
[tool.poetry]
name = "learn-fastapi"
version = "0.1.0"
description = ""
authors = ["Lux"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-dotenv = "0.21.0"
validators = "0.18.2"
sqlalchemy = "2.0.23"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
psycopg = "^3.1.13"
psycopg-binary = "^3.1.13"
aiosqlite = "0.20.0"
pyjwt = "^2.9.0"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
python-multipart = "^0.0.9"
environs = "^11.0.0"
snowflake-id = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
